requirements.txt
# Discord
discord.py

# Google SDK - use the latest available version
google-adk
google-generativeai

# Database
qdrant-client
mem0ai

# Web scraping and search
aiohttp==3.9.1
newspaper3k==0.2.8
nltk==3.8.1
lxml[html_clean]
lxml
lxml_html_clean

# Utility libraries
python-dotenv==1.0.0
